implemented predict and release version methods in version api - switched to version api in core package - moved all remaining 'rogue' git executions to the git api
