+++ pkg/core/get.go
@@ -1,7 +1,7 @@
 package core
 
 import (
-	"github.com/restechnica/semverbot/pkg/version"
+	"github.com/restechnica/semverbot/pkg/versions"
 )
 
 type GetVersionOptions struct {
@@ -11,6 +11,6 @@ type GetVersionOptions struct {
 // GetVersion gets the current version based on the latest annotated git tag.
 // Returns the current version.
 func GetVersion(options *GetVersionOptions) string {
-	var versionAPI = version.NewAPI()
+	var versionAPI = versions.NewAPI()
 	return versionAPI.GetVersionOrDefault(options.DefaultVersion)
 }
