+++ pkg/core/push.go
@@ -1,10 +1,7 @@
 package core
 
 import (
-	"fmt"
-
-	"github.com/restechnica/semverbot/pkg/git"
-	"github.com/restechnica/semverbot/pkg/version"
+	"github.com/restechnica/semverbot/pkg/versions"
 )
 
 type PushVersionOptions struct {
@@ -15,11 +12,7 @@ type PushVersionOptions struct {
 // PushVersion pushes the latest annotated git tag to the git origin.
 // Returns an error if pushing the tag went wrong.
 func PushVersion(options *PushVersionOptions) (err error) {
-	var versionAPI = version.NewAPI()
+	var versionAPI = versions.NewAPI()
 	var version = versionAPI.GetVersionOrDefault(options.DefaultVersion)
-
-	var prefixedVersion = fmt.Sprintf("%s%s", options.GitTagsPrefix, version)
-
-	var gitAPI = git.NewAPI()
-	return gitAPI.PushTag(prefixedVersion)
+	return versionAPI.PushVersion(version, options.GitTagsPrefix)
 }
