+++ pkg/versions/prefix.go
@@ -1,0 +1,7 @@
+package versions
+
+import "fmt"
+
+func AddPrefix(version string, prefix string) string {
+	return fmt.Sprintf("%s%s", prefix, version)
+}
